Rework of admin Dashboadrd (1) use generic forms for edit / show / create compabitlity with coreui v3

+++ app/Http/Controllers/Admin/BugreportsCommentController.php
@@ -13,16 +13,19 @@ class BugreportsCommentController extends Controller
     public function store(StoreBugreportCommentRequest $request)
     {
         abort_unless(\Gate::allows('bugreportComment_create'), 403);
+        abort_unless(\Auth::user()->id == $request->user_id, 403);
 
         $bugreportComment = BugreportComment::create($request->all());
+        $bugreportComment->user_id = \Auth::user()->id;
 
         return redirect()->route('admin.bugreports.show', [$bugreportComment->bugreport_id]);
     }
 
-    public function update(UpdateBugreportCommentRequest $request, BugreportComment $bugreportComment)
+    public function update(UpdateBugreportCommentRequest $request)
     {
         abort_unless(\Gate::allows('bugreportComment_edit'), 403);
         $bugreportComment = BugreportComment::find($request->get('id'));
+        abort_unless(\Auth::user()->id == $bugreportComment->user_id, 403);
         $bugreportComment->update($request->all());
 
         return redirect()->route('admin.bugreports.show', [$bugreportComment->bugreport_id]);
@@ -31,6 +34,7 @@ class BugreportsCommentController extends Controller
     public function destroy($id)
     {
         abort_unless(\Gate::allows('bugreportComment_delete'), 403);
+        abort_unless(\Auth::user()->id == BugreportComment::find($id)->user_id, 403);
 
         BugreportComment::find($id)->delete();
